// src/payload-types.ts
export interface Media {
  id: number;
  /**
   * If left empty, this will be auto-populated from the filename.
   */
  alt?: string | null;
  /**
   * Categorizing media helps in organizing the library and Cloudinary folders.
   */
  category?:
    | (
        | 'ypsl'
        | 'executive-committees'
        | 'standing-committees'
        | 'ai-driven-sri-lanka'
        | 'sl-inspire'
        | 'lets-talk'
        | 'insl'
        | 'studpro'
        | 'y2npro'
        | 'merch-payslips'
        | 'others'
        | 'merchants'
        | 'ieee-projects'
        | 'events'
      )
    | null;
  cloudinaryUrl?: string | null;
  cloudinaryPublicId?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface Merchant {
  id: number;
  /**
   * A unique identifier for this item (e.g., YPSL-TSHIRT-001)
   */
  merchantID: string;
  merchantName: string;
  category: number | MerchCategory;
  price: number;
  discountTag?: string | null;
  statusTag: 'available' | 'out-of-stock' | 'limited';
  availability: 'available' | 'closed';
  sizes?: ('XS' | 'S' | 'M' | 'L' | 'XL' | 'XXL' | 'XXXL' | 'Free Size')[] | null;
  images: {
    image: number | Media;
    id?: string | null;
  }[];
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "merch-categories".
 */
export interface MerchCategory {
  id: number;
  name: string;
  /**
   * Used for filtering (e.g., t-shirt, hoodie)
   */
  slug: string;
  updatedAt: string;
  createdAt: string;
}